Reduce physics update jitter
Added Main::target_iters, modifying physics iteration exit condition
and target_iters so that the number of physics frame per render frame
tries to be between target_iters and target_iters+1. Reduces jitter
when the physics framerate almost, but not quite, matches the render
framerate or an integer multiple thereof.