Add OS.max_pending_frames parameter
to control the latency vs. througput behavior of syncing the CPU to
GPU swaps. Also available as project setting
rendering/threads/max_pending_frames. The definition is: before input
is processed, it is made sure that no more than max_pending_frames of
past rendering frames are still being processed.

If max_pending frames is set to -1, the new code in this commit has no effect.

The purpose is twofold. The main motivation is that it enables
fetching timer information via glQueryCounter from the time the
graphics card actually swaps the buffers or is otherwise done
rendering a frame. To get stable timings this way, however, one needs
to fix the number of frames to go back to fetch the timing and you
need to wait for the frame in question to actually finish rendering
(and swapping). OS.max_pending_frames controls precisely that.

Lower values of OS.max_pending_frames give lower input latency
measured in frames, higher values yield better througput in frames per
second.

Values in particular:
-1: Do not wait on GPU, do not take timings from GPU.
0 : Before processing input for the next frame, wait for the last
    frame to finish swapping. This effectively disables any parallelism
    between the main thread and the rendering thread and limits
    parallelism between GPU and CPU. Use only for the simplest games
    and if low latency is very important.
1 : Before the visual server begins work on a a new frame, wait for the last
    frame to finish swapping. This does not limit parallelism between
    main and render thread, but still limits GPU use. Use if low
    latency is important and you can be reasonably sure the game will
    not be GPU limited.
2 : Before the visual server begins work on a a new frame, wait for
    the next to last frame to finish swapping. Puts no limits on
    parallelism if there is only one GPU. This is a safe choice.
3 : Wait for frames even further back. Tangible benefits in througput
    over 2 can only be expected for multi-GPU setups.