Collecting frame end time with glQueryCounter.
Collection is triggered in rasterizer_gles3.cpp in
RasterizerGLES3::end_frame(). The new virtual function
Rasterizer::sync_end_frame() can be used to later fetch the result.
The virtual function VisualServer::sync() and its implementations have
been adapted to pass the result out to the main loop.

Side effect: Rasterizer::sync_end_frame() and therefore
VisualServer::sync() will now actively wait for the next-to-last frame
to finish rendering. The last frame can still be fully pending in the
GPU/Driver pipleline, so this should have no impact on throughput in
normal situations. It does increase the risk of frame drops if the CPU
has unusually much to do for a frame or two and additional bufferd
frames could compensate, so the precise frame to wait for (or whether
there should be any waiting) should be made configurable later.