Timer evaluation refactoring: created _TimerSyncClassic
The new class handles the decisions about simulation progression
previously handled by main::iteration(). It is fed the current timer
ticks and determines how many physics updates are to be run and what
the delta argument to the _process() functions should be.

Clamping the updates to 8 physics updates max is left in
main::iteration().

No functional change yet.