Select between CPU and GPU based timers based on their stability
Keep track of past frame times reported by the main thread CPU timer
and the GPU timer listening to buffer swaps. Pick the consistently
better one. Even if the GPU timer is doing its job perfectly and
giving us the exact times a buffer became visible, the CPU timer may
be the better choice for a smooth experience, for example if the game
runs choppily because of high CPU load.

Keep track of timer drift and compensate by clamping to the main CPU timer.