Keep physics steps consistent over multiple frames
To avoid stuttering when the physics FPS does not match the screen
refresh at all (60 FPS physics on a 144 or 120 Hz screen), we need to
keep track of the typical number of physics steps accumulated over the
last N frames; otherwise, the same logic as before with single frame
tracking applies. The next accumulated update count should be a stored
typical value for it or one higher. Changes to the typical value
happen only if a threshold is surpassed.