Add parameter physics_steps_change_threshold
Threshold controlling the physics step count calculation to achieve
smoother updates. It determines how hard it is to change the physics
step count from one render frame to the next. At 0, there is no
resistance to change. Valid values are between 0 and 1. At 1, the
system will not go from 1 steps per frame to 2 unless it would
normally have gone to 3 steps. Sensible values are in the range of
0.01 and 0.1.