clamping animation timestep and indirectly smoothing it
Based on the information gathered from physics updates, clamp the
animation timesteps so they would be compatible with the physics
updates if they were constant. Since one bound on the timesteps gained
this way usually is pretty close to the actual average timestep, this
is quite good at removing small jitters.