Add measurement of timer variance
Principle: Past frame times and past time deficits are
recorded. Separately, min and max are determined of them,
the difference is a variance measure. Take the max of the two,
and also let it not shrink too much from frame to frame because
that causes trouble in the compensation code.