Heavily rework MainTimerSync
The basic principle stays the same. From the input timesteps,
we first try to extract an appropriate rhythm of physics steps
for a sequence of game loop runs.

Previously, that rhythm was both calculated using and applied
to updates with the configured physics_jitter_fix.
Now, we calculate the rhythm based on a fixed physics_jitter_fix
of 0.5 because that gives the most stable result. That rhythm is
then used as a base for the regular update process using the
configured jitter fix.

The source location where each of the many clamping steps is done
is shuffled around to hopefully make them more understandable.